pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "fill"
version = "0.1.0"
description = "Jinja-based batch SQL generator with batch-sql preprocessor"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.8"
dependencies = [
    "Jinja2>=3.1.0,<4",
    "PyYAML>=6.0"
]

[project.scripts]
fill = "fill.cli:main"
